requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Ingestion
feedparser==6.0.10
arxiv==2.0.0
requests==2.31.0
beautifulsoup4==4.12.3

# Embeddings & Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# RAG & LLM (versions compatibles)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
litellm>=1.45.0
groq>=0.4.0

# CLI Interface
typer>=0.24.0
rich==13.7.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
